4:44PM

+++ api.py
@@ -13,8 +13,9 @@
 #1. Laika noteiksana latvija
 import requests
 import json
-responseASV = requests.get("http://worldtimeapi.org/api/timezone/America/New_York")
-responseRiga = requests.get("http://worldtimeapi.org/api/timezone/Europe/Riga")
+import os
+response1 = requests.get("http://worldtimeapi.org/api/timezone/Europe/Riga")
+response2 = requests.get("http://worldtimeapi.org/api/timezone/Asia/Tokyo")
 
 def getLocation(obj):
     location = obj.json()["timezone"]
@@ -29,7 +30,7 @@ def getTime(obj):
     return laiks
 
 def laikaStarpiba(timezone1,timezone2):
-    starpiba = abs(int((getTimezone(timezone1))) - int(getTimezone(timezone2)))
+    starpiba = int((getTimezone(timezone2))) - int(getTimezone(timezone1))
     return starpiba
 
 def DataPreparation(obj1,obj2):
@@ -37,12 +38,23 @@ def DataPreparation(obj1,obj2):
         f"Laiks {getLocation(obj1)}": getTime(obj1),
         f"Laiks {getLocation(obj2)}": getTime(obj2),
         f"Laika_Zona {getLocation(obj1)}": getTimezone(obj1),
-        f"Laiks_Zona {getLocation(obj2)}": getTimezone(obj2),
+        f"Laika_Zona {getLocation(obj2)}": getTimezone(obj2),
         f"Laika_Starpiba {getLocation(obj1),getLocation(obj2)}": f"{laikaStarpiba(obj1,obj2)} h"
     }
     return dati
 
-json_object = json.dumps(DataPreparation(responseASV,responseRiga), indent=4)
-
-with open("laiksAPI.json", "w") as file:
-    file.write(json_object)
+if not os.path.exists("laiksAPI.json"):
+    with open("laiksAPI.json", "w") as file:
+        data1 = [DataPreparation(response1,response2)]
+        json_object = json.dumps(data1, indent=4)
+        file.write(json_object)
+else:
+    oldData = []
+    with open("laiksAPI.json", "r+") as file:
+        oldData = json.load(file)
+        newData = json.loads(json.dumps(DataPreparation(response1,response2), indent=4))
+        oldData.append(newData)
+    with open("laiksAPI.json", "w") as file:
+        file.write(json.dumps(oldData, indent=4))
+        
+
